create a middleware helper file
