+++ modules/din_middleware_test.go
@@ -9,8 +9,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/DIN-center/din-sc/apps/din-go/lib/din"
-
 	prom "github.com/DIN-center/din-caddy-plugins/lib/prometheus"
 	"github.com/caddyserver/caddy/v2"
 	"github.com/caddyserver/caddy/v2/caddyconfig/caddyfile"
@@ -345,87 +343,6 @@ func TestUnmarshalCaddyfile(t *testing.T) {
 			err := dinMiddleware.UnmarshalCaddyfile(dispenser)
 			if err != nil && !tt.hasErr {
 				t.Errorf("UnmarshalCaddyfile() = %v, want %v", err, tt.hasErr)
-			}
-		})
-	}
-}
-
-func TestSyncRegistryWithLatestBlock(t *testing.T) {
-	logger := zap.NewNop()
-	mockCtrl := gomock.NewController(t)
-	mockDingoClient := din.NewMockIDingoClient(mockCtrl)
-	dinMiddleware := &DinMiddleware{
-		RegistryEnv:                         LineaMainnet,
-		RegistryBlockEpoch:                  10,
-		RegistryLastUpdatedEpochBlockNumber: 40,
-		logger:                              logger,
-		DingoClient:                         mockDingoClient,
-	}
-
-	tests := []struct {
-		name                                string
-		registryLastUpdatedEpochBlockNumber int64
-		latestBlockNumber                   int64
-		expectedUpdateCall                  bool
-		expectedBlockFloorByEpoch           int64
-	}{
-		{
-			name:                                "Sync should update as block difference is equal to or exceeds epoch 50",
-			registryLastUpdatedEpochBlockNumber: int64(40),
-			latestBlockNumber:                   int64(50),
-			expectedUpdateCall:                  true,
-			expectedBlockFloorByEpoch:           int64(50),
-		},
-		{
-			name:                                "Sync should update as block difference is equal to or exceeds epoch 52",
-			registryLastUpdatedEpochBlockNumber: int64(40),
-			latestBlockNumber:                   int64(52),
-			expectedUpdateCall:                  true,
-			expectedBlockFloorByEpoch:           int64(50),
-		},
-		{
-			name:                                "Sync should update as block difference is equal to or exceeds epoch 1000",
-			registryLastUpdatedEpochBlockNumber: int64(40),
-			latestBlockNumber:                   int64(1001),
-			expectedUpdateCall:                  true,
-			expectedBlockFloorByEpoch:           int64(1000),
-		},
-		{
-			name:                                "Sync should not update as block difference is less than epoch 48",
-			registryLastUpdatedEpochBlockNumber: int64(40),
-			latestBlockNumber:                   int64(48),
-			expectedUpdateCall:                  false,
-			expectedBlockFloorByEpoch:           int64(40),
-		},
-		{
-			name:                                "Sync should not update as block difference is less than epoch 30",
-			registryLastUpdatedEpochBlockNumber: int64(40),
-			latestBlockNumber:                   int64(30),
-			expectedUpdateCall:                  false,
-			expectedBlockFloorByEpoch:           int64(40),
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			dinMiddleware.Networks = map[string]*network{
-				LineaMainnet: {
-					LatestBlockNumber: tt.latestBlockNumber,
-				},
-			}
-			dinMiddleware.RegistryLastUpdatedEpochBlockNumber = tt.registryLastUpdatedEpochBlockNumber
-
-			// Check if update was called as expected
-			if tt.expectedUpdateCall {
-				mockDingoClient.EXPECT().GetRegistryData().Return(&din.DinRegistryData{}, nil).Times(1)
-			}
-			// Call the function
-			dinMiddleware.syncRegistryWithLatestBlock()
-
-			// Validate that RegistryLastUpdatedEpochBlockNumber is updated correctly
-			if dinMiddleware.RegistryLastUpdatedEpochBlockNumber != tt.expectedBlockFloorByEpoch {
-				t.Errorf("Expected RegistryLastUpdatedEpochBlockNumber = %v, got %v", tt.expectedBlockFloorByEpoch, dinMiddleware.RegistryLastUpdatedEpochBlockNumber)
 			}
 		})
 	}
